Reorganized build.
The maven reactor is now reordered for proper sequencing for building isoimage in l4 as our final artifact.
Next job is to configure l4 to grab the dependencies and do it's thing with them. then ewe have to modify modules.list and jvm_server.cfg and main.cc to have a bootable java -version and die working.